+++ forui/lib/widgets/select.dart
@@ -12,6 +12,8 @@ export '../src/widgets/select/content/scroll_handle.dart' hide ScrollHandle;
 export '../src/widgets/select/content/search_content.dart' hide SearchContent;
 export '../src/widgets/select/content/search_field_properties.dart';
 export '../src/widgets/select/multi/select.dart';
+export '../src/widgets/select/multi/select_style.dart';
 export '../src/widgets/select/multi/tag.dart';
 export '../src/widgets/select/single/select.dart';
+export '../src/widgets/select/single/select_style.dart';
 export '../src/widgets/select/single/select_controller.dart' hide InternalFSelectControl;